philos lives (sleeping, eating, thinking, dying).

+++ philo_one/src/libft/libft.h
@@ -34,7 +34,7 @@ void				*ft_memmove(void *dst, const void *src, size_t len);
 void				*ft_memset(void *b, int c, size_t len);
 void				ft_putchar_fd(char c, int fd);
 void				ft_putendl_fd(char *s, int fd);
-void				ft_putnbr_fd(int n, int fd);
+void				ft_putnbr_fd(uint64_t n, int fd);
 int					ft_nbcmp(const char *nb1, const char *nb2);
 void				ft_putstr_fd(char *s, int fd);
 char				**ft_split(char const *s, char c);
